Site redesign
I have redesigned the site to take advantage of more of the Bootstrap
features, including the responsive layout, and to have a cleaner look
and feel. Not all of the site's features have been updated for the new
design and there will likely be bugs, as you all were committing while I
was making the redesign, but that will come in time. I would like
feedback. If you don't like it you can always revert this commit.